# docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
...
  httpx.JSendError:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: Failed to connect to database
        type: string
      status:
        example: error
        type: string
    type: object
  httpx.JSendFail:
    properties:
      data:
        additionalProperties: {}
        type: object
      status:
        example: fail
        type: string
    type: object
...
  models.UserProfile:
    properties:
      created_at:
        example: "2025-01-15T10:30:00Z"
        type: string
      email:
        example: juan.perez@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Juan Pérez
        type: string
      phone:
        example: "+525512345678"
        type: string
      role:
        enum:
        - customer
        - merchant
        - driver
        - admin
        example: customer
        type: string
      updated_at:
        example: "2025-01-15T10:30:00Z"
        type: string
    type: object
  models.UserProfileResponse:
    properties:
      data:
        $ref: '#/definitions/models.UserProfile'
      status:
        example: success
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Delivery marketplace API for customers, merchants, and drivers.
  title: TacoShare Delivery API
  version: "1.0"
paths:
...
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get a user's profile information by their UUID. Users can only
        access their own profile unless they have admin role. Returns full user details
        including ID, name, email, phone, role, and timestamps.
      parameters:
      - description: User ID in UUID format (e.g., 550e8400-e29b-41d4-a716-446655440000)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User profile retrieved successfully with all user details
          schema:
            $ref: '#/definitions/models.UserProfileResponse'
        "400":
          description: Invalid user ID format - must be valid UUID
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "403":
          description: Forbidden - insufficient permissions to access this user's
            profile
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "404":
          description: User not found - no user exists with the provided ID
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "500":
          description: Internal server error - failed to retrieve user profile
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
  /users/me:
    get:
      consumes:
      - application/json
      description: Get the authenticated user's profile information including ID,
        name, email, phone, role, and timestamps. Requires valid JWT token in Authorization
        header.
      produces:
      - application/json
      responses:
        "200":
          description: User profile retrieved successfully with all user details
          schema:
            $ref: '#/definitions/models.UserProfileResponse'
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "404":
          description: User not found - user ID from token doesn't exist in database
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "500":
          description: Internal server error - failed to retrieve user profile
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
produces:
- application/json
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"